backend/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
beautifulsoup4
requests
pypdf
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
